move cluster configuration
